create, edit, delete, list refactoring

+++ src/main/java/kz/trei/acs/action/account/DeleteUser.java
@@ -6,7 +6,7 @@ import kz.trei.acs.action.ActionType;
 import kz.trei.acs.dao.DaoException;
 import kz.trei.acs.dao.DaoFactory;
 import kz.trei.acs.dao.UserDao;
-import kz.trei.acs.util.PropertyManager;
+import kz.trei.acs.util.GetParameterException;
 import org.apache.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
@@ -15,33 +15,30 @@ import javax.servlet.http.HttpServletResponse;
 public class DeleteUser implements Action {
     private static final Logger LOGGER = Logger.getLogger(DeleteUser.class);
 
-    static {
-        PropertyManager.load("configure.properties");
-    }
-
     @Override
     public ActionResult execute(HttpServletRequest request, HttpServletResponse response) {
+        LOGGER.debug("...");
         response.setCharacterEncoding("UTF-8");
         long id;
-        try {
-            id = Long.valueOf(request.getParameter("id"));
-        } catch (NumberFormatException e) {
-            LOGGER.error("Wrong id parameter " + e.getMessage());
-            return new ActionResult(ActionType.REDIRECT, "user-list?status=id.parameter.error");
-        }
         DaoFactory daoFactory = DaoFactory.getFactory();
         UserDao userDao = daoFactory.getUserDao();
-        if (id == 1) {
-            LOGGER.debug("was trying to delete the read-only record id==1");
-            return new ActionResult(ActionType.REDIRECT, "user-list?status=delete.user.fail");
-        }
         try {
+            id = UserUtil.takeIdFromRequest(request);
+            if (id == 0 || id == 1) {
+                throw new GetParameterException("trying to delete the read-only record");
+            }
             userDao.delete(id);
         } catch (DaoException e) {
+            request.setAttribute("status", "delete.user.fail");
             LOGGER.error("DAO delete error " + e.getMessage());
-            return new ActionResult(ActionType.REDIRECT, "user-list?status=delete.user.fail");
+            return new ActionResult(ActionType.REDIRECT, "user-list" + UserUtil.fetchParameters(request));
+        } catch (GetParameterException e) {
+            request.setAttribute("status", "error.parameter.id.invalid");
+            LOGGER.error(e.getMessage());
+            return new ActionResult(ActionType.REDIRECT, "user-list" + UserUtil.fetchParameters(request));
         }
+        request.setAttribute("status", "delete.user.success");
         LOGGER.debug("The user is deleted successfully");
-        return new ActionResult(ActionType.REDIRECT, "user-list?status=delete.user.success");
+        return new ActionResult(ActionType.REDIRECT, "user-list" + UserUtil.fetchParameters(request));
     }
 }
