+++ src/main/java/kz/trei/acs/action/rfidtag/DeleteRfidTag.java
@@ -6,7 +6,7 @@ import kz.trei.acs.action.ActionType;
 import kz.trei.acs.dao.DaoException;
 import kz.trei.acs.dao.DaoFactory;
 import kz.trei.acs.dao.RfidTagDao;
-import kz.trei.acs.util.PropertyManager;
+import kz.trei.acs.util.GetParameterException;
 import org.apache.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
@@ -16,29 +16,30 @@ import javax.servlet.http.HttpServletResponse;
 public class DeleteRfidTag implements Action {
     private static final Logger LOGGER = Logger.getLogger(DeleteRfidTag.class);
 
-    static {
-        PropertyManager.load("configure.properties");
-    }
-
     @Override
     public ActionResult execute(HttpServletRequest request, HttpServletResponse response) {
+        LOGGER.debug("...");
         response.setCharacterEncoding("UTF-8");
         long id;
-        try {
-            id = Long.valueOf(request.getParameter("id"));
-        } catch (NumberFormatException e) {
-            LOGGER.error("Wrong id parameter " + e.getMessage());
-            return new ActionResult(ActionType.REDIRECT, "rfidtag-list?status=id.parameter.error");
-        }
         DaoFactory daoFactory = DaoFactory.getFactory();
         RfidTagDao rfidTagDao = daoFactory.getRfidTagDao();
         try {
+            id = RfidTagUtil.takeIdFromRequest(request);
+            if (id == 0 || id == 1) {
+                throw new GetParameterException("trying to delete the read-only record");
+            }
             rfidTagDao.delete(id);
         } catch (DaoException e) {
+            request.setAttribute("status", "delete.rfidtag.fail");
             LOGGER.error("DAO delete error " + e.getMessage());
-            return new ActionResult(ActionType.REDIRECT, "rfidtag-list?status=delete.rfidtag.fail");
+            return new ActionResult(ActionType.REDIRECT, "rfidtag-list" + RfidTagUtil.fetchParameters(request));
+        } catch (GetParameterException e) {
+            request.setAttribute("status", "error.parameter.id.invalid");
+            LOGGER.error(e.getMessage());
+            return new ActionResult(ActionType.REDIRECT, "rfidtag-list" + RfidTagUtil.fetchParameters(request));
         }
-        LOGGER.debug("The RFID tag is deleted successfully");
-        return new ActionResult(ActionType.REDIRECT, "rfidtag-list?status=delete.rfidtag.success");
+        request.setAttribute("status", "delete.rfidtag.success");
+        LOGGER.debug("The rfidtag is deleted successfully");
+        return new ActionResult(ActionType.REDIRECT, "rfidtag-list" + RfidTagUtil.fetchParameters(request));
     }
 }
